% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{betriebsergebnis_speichern}
\alias{betriebsergebnis_speichern}
\title{Reads ADED entity 880028}
\usage{
betriebsergebnis_speichern(header_list)
}
\arguments{
\item{header_list}{A list as returned by \code{\link[RAdis]{read_adis}} containing
the ADIS entities of a single header.}
}
\value{
A data frame with the herd results of the dairy herd improvement tests.
}
\description{
Used inside \code{\link{prepare_PCstart}}.
}

