% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_functions.R
\name{calculate_indicator}
\alias{calculate_indicator}
\title{Calculations of dairy herd health indicators from Dairy Herd Improvement Test data}
\usage{
calculate_indicator(indicator, data, testmonths = NA)
}
\arguments{
\item{indicator}{A character string, one of \code{
c("LowSCC", "LactNI", "LactCure", "Chronic", "NoCure", "DryNI", "DryCure", "HeiferMast",
"LowFPR", "HiFPRtotal", "HiFPR100", "Urea")}}

\item{data}{A dataframe like \code{$einzeltiere} returned by \code{\link{prepare_PCstart}}.}

\item{testmonths}{The months for which the indicator should be calculated as vector
of class \code{Date}. If \code{NA}, the recent date of \code{data} is used.}
}
\description{
Used inside \code{\link{IndicatorSheet}}.
}
\references{
Visit \url{http://www.milchqplus.de/kennzahlen_1.html} for descriptions
 of the udder health indicators.
}

