% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{einzeltierliste_erstellen}
\alias{einzeltierliste_erstellen}
\title{Used inside \link{monitor_SCCUdderHealth}.}
\usage{
einzeltierliste_erstellen(mlpdaten, kuehe, monate = 3)
}
\description{
Used inside \link{monitor_SCCUdderHealth}.
}

