% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_functions.R
\name{IndicatorSheet}
\alias{IndicatorSheet}
\title{Create a simple graphical report of the most important SCC indicators}
\usage{
IndicatorSheet(PCstart = NULL, Farm = "SomeDairyHerd", out_dir = NULL,
  out_window = FALSE, return_results = FALSE)
}
\arguments{
\item{PCstart}{Character of length 1. Full path to the ADIS file to be used for monitoring.}

\item{Farm}{A character string, the 'name' used for the herd / farm in the report.}
}
\value{
_results Logical of length 1. Should the data shown in the report be returned
 as a data.frame? (Default is \code{FALSE}).

A grafical report in pdf-format is always produced in the \code{out_dir}
 directory. If \code{return_results} is \code{TRUE}, the data is also returned
 as data.frame. \code{NULL} invisible otherwise.
}
\description{
This function calculates the most important udder health indicators that are
based on the monthly dairy herd improvement test somatic cell counts (SCC)
according to the DLQ and \url{http://www.milchqplus.de} and produces a concise
graphical report on a single page in pdf-format.
}
\details{
See \code{\link{prepare_PCstart}} for the required input file.
}
\note{
Currently different 'AE's are ignored and all animals of one herd are analysed together.

This function does not require LaTeX.
}
\references{
Visit \url{http://www.milchqplus.de} for further descriptions
 of the udder health indicators.

Have a look at \cite{V. Zoche, W. Heuwieser, and V. Kroemker, "Risk-based monitoring of udder health. A review."}, \href{http://www.schattauer.de/t3page/1214.html?manuscript=16040}{Tierärztliche Praxis. Ausgabe G, Grosstiere/Nutztiere, vol. 39, no. 2, pp. 88–94, 2011.} and \href{http://www.lkvsachsen.de/fileadmin/lkv/redaktion/download/unternehmen/Veredlungsland/Handout_Euterschulung.pdf}{these slides} for information about monitoring the udder health.
}

