% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main_functions.R
\name{monitor_SCCUdderHealth}
\alias{monitor_SCCUdderHealth}
\title{The main function of UdderHealthMonitor.}
\usage{
monitor_SCCUdderHealth(PCstart = NULL, Betriebsname = NULL,
  Analyst = NULL, return_results = FALSE,
  settings = settings_SCCmonitor(), ...)
}
\arguments{
\item{PCstart}{Character of length 1. Full path to the PCStart ADIS file to be used for monitoring, expected in UNIX-style ('/').}

\item{return_results}{If \code{TRUE} the dataframes ds_monatlich and ds_12Monate are return in a list.}

\item{settings}{List. See the documentation of \link{settings_SCCmonitor} for details.}

\item{...}{Further arguments passed to inner and report functions (e.g.
 \code{generate_report(usertemplate)}).}
}
\description{
The main function of UdderHealthMonitor.
}

