% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{prepare_PCstart}
\alias{prepare_PCstart}
\title{Prepares the data of the 'PCstart' ADIS files from VIT for udder health monitoring.}
\usage{
prepare_PCstart(file, ...)
}
\arguments{
\item{file}{Character of length 1. Full path to the PCStart ADIS file to be read, expected in UNIX-style ('/').}

\item{...}{Further arguments passed to \link{read_adis}.}
}
\description{
Used inside \link{monitor_SCCUdderHealth}.
}

