% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{prepare_SCCdata}
\alias{prepare_SCCdata}
\title{Prepares individual cows' data for SCC analysis.}
\usage{
prepare_SCCdata(inddata, Months = c(1:12))
}
\arguments{
\item{inddata}{A dataframe provided as \code{$einzeltiere} by \link{prepare_PCstart}.}

\item{Months}{An integer vector giving the months for which the data should be prepared:
1 is the most recent month, 2 the second most recent month, ... (default = 1:12).}
}
\description{
Used inside \link{monitor_SCCUdderHealth}.
}

