% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main_functions.R
\name{settings_SCCmonitor}
\alias{settings_SCCmonitor}
\title{Settings for monitor_SCCUdderHealth.}
\usage{
settings_SCCmonitor()
}
\description{
\code{$GW[1]} SCC cut-off for Neuinfektionen, Heilung, Eutergesund
\code{$GW[2]} SCC cut-off for Chronisch, (Regression -> erweiterungen)
\code{$GW[3]} SCC cut-off for Unheilbar
}

