% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/report_functions.R
\name{cleanKnit2pdf}
\alias{cleanKnit2pdf}
\title{Converts Rnw files to PDF using \link{knitr::knit2pdf} without leaving any other files.}
\usage{
cleanKnit2pdf(infile, outfile, useEnvir)
}
\arguments{
\item{infile}{A string defining the output file with complete path.}

\item{outfile}{A string defining the output file with complete path.}

\item{useEnvir}{The environment in which the code chunks are to be evaluated
 by \code{knit2pdf()}.}
}
\description{
Converts Rnw files to PDF using \link{knitr::knit2pdf} without leaving any other files.
}

