% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/report_functions.R
\name{generate_report}
\alias{generate_report}
\title{Generates the UdderHealthReport using a Sweave-template.}
\usage{
generate_report(outfile, useEnvir, usertemplate = NULL, replace = NULL)
}
\arguments{
\item{outfile}{A string defining the output file with complete path.}

\item{useEnvir}{The environment in which the code chunks are to be evaluated
by \code{knit2pdf()}.}

\item{usertemplate}{An appropriate Sweave-File containing the template for the report.
See 'UdderHealthReport.Rnw' as example. If \code{NULL}, the default template from
'UdderHealthReport.Rnw' is used.}

\item{replace}{A named character vector containing the replacements for the
 wildcards (\code{|TMPwildcard{name}|}) in the template. The names of \code{replace}
 are searched for in the wildcard names.}
}
\description{
Generates the UdderHealthReport using a Sweave-template.
}

