% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/report_functions.R
\name{generate_template}
\alias{generate_template}
\title{Generates a reporting template for \code{generate_report()} from a Sweave-file.}
\usage{
generate_template(template, output = "text", ...)
}
\arguments{
\item{template}{Sweave-file (.Rnw), the template to be used}

\item{output}{One of \code{c("text", "character")}}

\item{...}{Further arguments passed to \code{readLines()}}
}
\value{
For \code{output = "text"} prints the template notated as string on the console.
 Then the output can be copied and pasted into a function’s code
 (for generating the packages default).
 For \code{output = "character"}, the result of \code{readLines(template)} is returned.
}
\description{
Generates a reporting template for \code{generate_report()} from a Sweave-file.
}
\note{
Do not use \code{''} - Quotes in the template but \code{""} instead.
}

