% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_functions.R
\name{IndicatorSheet}
\alias{IndicatorSheet}
\title{Create a simple graphical report of the most important SCC indicators}
\usage{
IndicatorSheet(PCstart = NULL, Farm = "SomeDairyHerd", out_dir = NULL,
  out_window = FALSE, return_results = FALSE)
}
\arguments{
\item{PCstart}{Character of length 1. Full path to the ADIS file to be used for monitoring.}

\item{Farm}{A character string, the 'name' used for the herd / farm in the report.}

\item{out_dir}{Optional character of length 1. Full path to the directory where the report
file should be produced. If not specified the directory of \code{PCstart} is used.}

\item{out_window}{Logical of length 1. Should the report be shown in the standard plot display?
(Default is \code{FALSE})}

\item{return_results}{Logical of length 1. Should the data shown in the report be returned
as a data.frame? (Default is \code{FALSE}).}
}
\value{
A grafical report in pdf-format is always produced in the \code{out_dir}
 directory. If \env{return_results = TRUE}, the calculated values of the indicators
 are also returned as a data.frame with one row per month and the columns:
 \describe{
	 \item{\code{Month}}{Test month}
  \item{\code{LowSCC}}{Percentage of cows without mastitis}
  \item{\code{LactNI}}{Lactational new infection rate}
  \item{\code{LactCure}}{Lactational cure rate}
  \item{\code{Chronic}}{Percentage of cows with chronic mastitis}
  \item{\code{NoCure}}{Percentage of cows with low chances of cure}
  \item{\code{DryNI}}{Dry period new infection rate}
  \item{\code{DryCure}}{Dry period cure rate}
  \item{\code{HeiferMast}}{Heifer mastitis rate}
 }
 \code{NULL} invisible otherwise.
}
\description{
This function calculates the most important udder health indicators that are
based on the monthly dairy herd improvement test somatic cell counts (SCC)
according to the DLQ and \url{http://www.milchqplus.de} and produces a concise
graphical report on a single page in pdf-format.  
The indicators are calculated for the 13 recent months for which data is provided.
}
\details{
See \code{\link{prepare_PCstart}} for the required input file. The 
  definitions of the indicators are given in the package vignette
  \file{UdderHealthMonitor::User Guide}.
}
\note{
Currently different 'AE's are ignored and all animals of one herd are analysed together.
}
\references{
Visit \url{http://www.milchqplus.de} for further descriptions
 of the udder health indicators.

Have a look at \cite{V. Zoche, W. Heuwieser, and V. Kroemker, "Risk-based monitoring of udder health. A review."}, \href{http://www.schattauer.de/t3page/1214.html?manuscript=16040}{Tierärztliche Praxis. Ausgabe G, Grosstiere/Nutztiere, vol. 39, no. 2, pp. 88–94, 2011.} and \href{http://www.lkvsachsen.de/fileadmin/lkv/redaktion/download/unternehmen/Veredlungsland/Handout_Euterschulung.pdf}{these slides} for information about monitoring the udder health.
}

