% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{kalbungen_speichern}
\alias{kalbungen_speichern}
\title{Reads ADIS entity 880012.}
\usage{
kalbungen_speichern(header_list)
}
\arguments{
\item{header_list}{A list as returned by \link{read_adis} containing the ADIS entities of one single header.}
}
\description{
Used inside prepare_PCstart.
}

