% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{kennzahlen_bestimmen}
\alias{kennzahlen_bestimmen}
\title{Calculates the most important SCC udder health indicators for a given dataframe.}
\usage{
kennzahlen_bestimmen(daten, kennzahl, gw, relativ = TRUE)
}
\arguments{
\item{daten}{A dataframe as returned by \link{prepare_SCCdata}.}

\item{kennzahl}{A character string, one of
c("ErstMLPKuh", "ErstMLPFaerse", "FaersenMast", "TSNeuinf", "TSHeilung",
"LaktNeuinf", "LaktHeilung", "LaktChronisch", "Eutergesund", "Unheilbar")}

\item{gw}{An integer between 0 and 999 as SCC cut-off value to discriminate between healthy and inflamed udders.}

\item{relativ}{Logical. Should the relative value of the udder health indicator be given?}
}
\value{
The chosen udder health indicator value as scalar.
}
\description{
Used inside \link{monitor_SCCUdderHealth}.
}

