% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{tierdaten_speichern}
\alias{tierdaten_speichern}
\title{Reads ADIS entity 880005}
\usage{
tierdaten_speichern(header_list)
}
\arguments{
\item{header_list}{A list as returned by \code{\link[RAdis]{read_adis}} containing
 the ADIS entities of a single header.}
}
\value{
A data frame containing ID, Name, date of birth and sex for all animals.
}
\description{
Used inside \code{\link{prepare_PCstart}}.
}

